+++ krypto-core/src/androidMain/kotlin/krypto/SecureRandomJvm.kt
@@ -1,0 +1,15 @@
+package krypto
+
+import java.security.SecureRandom
+
+private val jrandom by lazy {
+    if (!randomUnittesting) {
+        PRNGFixes.apply()
+    }
+    SecureRandom()
+}
+
+actual fun fillRandomBytes(array: ByteArray) {
+    jrandom.nextBytes(array)
+}
+