+++ krypto-core/src/commonMain/kotlin/krypto/internal/KryptoTools.kt
@@ -1,0 +1,22 @@
+package krypto.internal
+
+internal inline fun Int.ext8(offset: Int) = (this ushr offset) and 0xFF
+
+internal fun Int.rotateRight(amount: Int): Int = (this ushr amount) or (this shl (32 - amount))
+internal fun Int.rotateLeft(bits: Int): Int = ((this shl bits) or (this ushr (32 - bits)))
+
+internal fun arraycopy(src: ByteArray, srcPos: Int, dst: ByteArray, dstPos: Int, count: Int) = src.copyInto(dst, dstPos, srcPos, srcPos + count)
+internal fun arraycopy(src: IntArray, srcPos: Int, dst: IntArray, dstPos: Int, count: Int) = src.copyInto(dst, dstPos, srcPos, srcPos + count)
+
+internal fun ByteArray.readU8(o: Int): Int = this[o].toInt() and 0xFF
+internal fun ByteArray.readS32_be(o: Int): Int =
+    (readU8(o + 3) shl 0) or (readU8(o + 2) shl 8) or (readU8(o + 1) shl 16) or (readU8(o + 0) shl 24)
+
+@PublishedApi internal fun min2(a: Int, b: Int) = if (a < b) a else b
+@PublishedApi internal fun max2(a: Int, b: Int) = if (a > b) a else b
+
+@PublishedApi internal fun min2(a: Float, b: Float) = if (a < b) a else b
+@PublishedApi internal fun max2(a: Float, b: Float) = if (a > b) a else b
+
+@PublishedApi internal fun min2(a: Double, b: Double) = if (a < b) a else b
+@PublishedApi internal fun max2(a: Double, b: Double) = if (a > b) a else b