+++ krypto-core/src/commonTest/kotlin/TestUtils.kt
@@ -1,0 +1,6 @@
+import krypto.encoding.Hex
+
+fun ByteArray.toHexString() = Hex.encode(this)
+fun ByteArray.toHexStringLower() = Hex.encodeLower(this)
+fun ByteArray.toHexStringUpper() = Hex.encodeUpper(this)
+val ByteArray.hex: String get() = toHexStringLower()