+++ krypto-keys/src/commonMain/kotlin/tz/co/asoft/KeyInfo.kt
@@ -1,7 +1,0 @@
-package tz.co.asoft
-
-interface KeyInfo {
-    var uid: String?
-    val created: Long
-    val expires: Long
-}