+++ CHANGELOG.md
@@ -1,26 +1,44 @@
+# 0.0.50 : 2021.06.02
+
+- Updated gradle from 6.7.1 to 6.8.1
+- Changed package from tz.co.asoft to krypto
+- Alligned with krypto master branch
+- Added support for native targets
+
 # 0.0.4 : 2020.12.12
+
 ## Optimizations
+
 - Removed test dependencies from commonMain
 
 # 0.0.3 : 2020.12.08
+
 ## Bug fixes
+
 - Added missing library configuration
 
 # 0.0.2 : 2020.12.08
+
 ## New Features
+
 - Added krypto-keys
 
 # 0.0.1 : 2020.11.27
+
 ## Availability
+
 - Published to maven central
 
 ## Build Src
+
 - Updated to gradle version 6.7.1
 
 ## Dependencies
+
 - Updated to kotlin 1.4.20
 
 ## New Features
+
 - Added reakt-buttons
 - Added reakt-feedbacks
 - Added reakt-icons
@@ -32,7 +50,9 @@
 - Added reakt-text
 
 ## Documentation
+
 - Update readme
 
 ## Samples
+
 - Added samples